Test chatroom
chatroom.sh - launches chatroomTest.py and sound.py alongside eachother.

chatroomtest.sh only launches chatroomTest so you can run soundTest alongside it.

soundTest takes functions from getAmplitudeTest and acts as the receiving script. The main idea is to run the script recursively so that the chatroom can constantly listen for input.

generatorTest remains as the sender and is called as a function in chatroomTest.